requirements.txt
PyJWT
requests
websockets
pydantic